Finish prototype of NextJS
